pyproject.toml
[tool.poetry]
name = "promptcheck"
version = "0.0.1a1"
description = "PromptCheck: CI for LLM Evals"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "promptcheck", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
typer = "^0.15.4"
pyyaml = "^6.0.2"
pydantic = "^2.11.4"
openai = "^1.78.1"
groq = "^0.24.0"
rouge-score = "^0.1.2"
tenacity = "^9.1.2"
nltk = {version = ">=3.9.1,<4.0.0", optional = true}

[tool.poetry.extras]
bleu = ["nltk"]

[tool.poetry.scripts]
promptcheck = "promptcheck.main:app"

[tool.poetry.group.dev.dependencies]
twine = "^6.1.0"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
